# parse_json._parse_location_history: speedup parsing about 50%