--- a/setup.cfg
+++ b/setup.cfg
@@ -49,6 +49,8 @@ console_scripts =
     google_takeout_parser = google_takeout_parser.__main__:main
 
 [options.extras_require]
+optional =
+    orjson
 testing =
     flake8
     mypy