- use orjson if it's available with fallback to builtin json
- use builtin fromisoformat from python 3.11, it's much faster (and simpler!)

- before:
  - parsing json 9.6s
  - processing data: 7.0s
- after
  - parsing json: 5.7s
  - processing data: 3.1s