@karlicoss 2024-09-11
mypy fails for the same reason I mentioned here https://github.com/seanbreckenridge/google_takeout_parser/pull/66#discussion_r1752921329 -- hopefully if we merge the other one and rebase it should be all good!